Package change from com.parmet to build.buf (#116)
This changes most of the `com.parmet` name spaced files to `build.buf`
in this repository. The remaining issue is the Buf CLI is hosted on
maven through `com.parmet` there is an internal Buf effort to publish
the CLI to maven. I'm planning on working with @pkwarren on this in the
coming days.