Update to Jackson 3.x and snakeyaml-engine (#359)
Users that are using snakeyaml (1.x or 2.x) may encounter issues with
the buf-gradle-plugin. Updating to Jackson 3.x changes the YAML support
to use snakeyaml-engine which doesn't have the same compatibility
issues, and this gets us on a newer baseline for YAML support in the
plugin.

The only caveat is that Jackson 3.x requires Java 17, so the plugin will
require that Gradle is run with Java 17+. This is similar to other
plugins in the ecosystem and Gradle 9 also requires Java 17.

Fixes #356.